created locationsReducers module and corrispetive test file
