+++ src/redux/reducers/reducers.js
@@ -1,40 +1,5 @@
 import * as types from '../actionTypes';
 
-export const locationReducer = (
-  state = {
-    loadingLocation: false,
-    locations: [],
-    allLocations: [],
-    error: null
-  },
-  action
-) => {
-  switch (action.type) {
-    case types.LOADING_LOCATIONS:
-      return { ...state, loadingLocation: true };
-    case types.FETCH_LOCATIONS_SUCCESS:
-      return {
-        ...state,
-        error: null,
-        loadingLocation: false,
-        locations: action.payload
-      };
-    case types.ALL_LOCATIONS_SUCCESS:
-      return {
-        ...state,
-        error: null,
-        loadingLocation: false,
-        allLocations: action.payload,
-        locations: action.payload
-      };
-    case types.FETCH_LOCATIONS_FAILURE:
-      return { ...state, loadingLocation: false, error: action.payload };
-    case types.CLEAR_LOCATIONS:
-      return { ...state, locations: [] };
-    default:
-      return state;
-  }
-};
 
 export const singleLocaReducer = (
   state = { loadingSingleLoc: false, location: null, error: null },
